[double-take] feat(double-take): initial commit (#1271)
* feat(double-take): initial commit

double-take is a unified UI and API for processing and training images for facial recognition. it
can be used in conjunction with frigate and home-assistant.

* docs(metadata): link appVersion to image tag

appVersion should link to the image tag to help the user understand what version of the application
is running

* docs(readme): regenerate helm docs

* docs(readme): warn user of configuration change implications

user should be notified that the configuration file will only be copied on the first run to a PVC,
at which point the config should be managed according to the upstream docs

* docs(readme): update config template to warn user of config management

updated the config template (in the correct place this time) so that it will warn the user about
making configuration changes. this should now persist when regenerating the helm-docs.