chore: fix unifi podsecurity context (#1611)
Fix Unifi pod `fsGroup`

When CSI mounts the unifi data if the `fsGroup` is not set it would be
mointed as uid `0` and when the unifi process starts as uid `999`, it
won't be able to access the data, producing permission denied errors and
the app failing to start.

Fix the above issue by setting the right `fsGroup` as set by the default
environment variables.

Signed-off-by: Noel Georgi <git@frezbo.dev>